transform/base64: Detect non-encoded content
Issue: 7114

Extend the `from_base64` transform with the `set_error` keyword. This
can be used to detect whether the input buffer is base64-encoded or not.

When `set_error` is used and the content is not encoded, the output
will be a buffer with fixed contents: `"BASE64_ECODE_BUF"`

Rule writers can use this with a following `content: "BASE64_ECODE_BUF"` to check
if the input buffer is not base64-encoded.

Add the content negation symbol `!` to ensure the content is
base64-encoded: `content: !"BASE64_ECODE_BUF"`