lua: update lua crate
Update lua crate to 0.1.0-alpha.4. This update will force a rewrite of
the headers if the env var SURICATA_LUA_SYS_HEADER_DST changes. This
fixes the issue where the headers may not be written.

The cause is that Rust dependencies are cached, and if your editor is
using rust-analyzer, it might cache the built without this var being
set, so these headers are not available to Suricata. This crate update
forces the re-run of the Lua build.rs if this env var changes, fixing
this issue.