lua: update lua crate with better header copying
Update lua crate to 0.1.0-alpha.5. This update will rewrite the
headers if the env var SURICATA_LUA_SYS_HEADER_DST changes. This fixes
the issue where the headers may not be written.

The cause is that Rust dependencies are cached, and if your editor is
using rust-analyzer, it might cache the built without this var being
set, so these headers are not available to Suricata. This crate update
forces the re-run of the Lua build.rs if this env var changes, fixing
this issue.

This new Lua crate is also smarter about writing out the headers,
don't overwrite if the destination appears to be the same as the
source, as this can force subsequent invocations of make to rebuild
the C source.