bindgen: use bindgen-cli instead of library
Replace bindgen usage from build.rs with the bindgen-cli program much
like we use cbindgen.

As bindgen can only accept one header file, we construct a
pseudo-header file of all the headers that need to be
consumed. Makefile dependency checking is done to make sure the
pseudo-header is only generated as needed to avoid rebuilding every
time.

Special handling is required for Windows to use the Windows path.