flow/manager: fix multi instance row tracking
In multi instance flow manager setups, each flow manager gets a slice
of the hash table to manage. Due to a logic error in the chunked
scanning of the hash slice, instances beyond the first would always
rescan the same subslice of their slice.

The `pos` variable that is used to keep the state of what the starting
position for the next scan was supposed to be was treated as if it held
a relative value. Relative to the bounds of the slice. It was however,
holding an absolute position. This meant that when doing it's bounds
check it was always considered out of bounds.

This patch addresses the issue by correctly handling the fact that the
value is absolute.

Bug: #7365.