datajson: add direct usage of JSON file
The format introduced in datajson is an evolution of the
historical datarep format. This has some limitations. For example,
if a user fetch IOCs from a threat intel server there is a large
change that the format will be JSON or XML. Suricata has no support
for the second but can support the first one.

This patch implements this concept. A optional json_key option is
added to the datajson keyword. If present, then Suricata assumes
that the data file contains a JSON object that is an array.
For each element elt of this array, the value added to the dataset
is elt['$json_key'] and the JSON data is elt itself.

Keeping the key value may seem redundant but it is useful to have it
directly accessible in the extra data to be able to query it
independantly of the signature (where it can be multiple metadata
or even be a transformed metadata).