flow: implement flood protection
This patch implements a TCP flood protection. It stores the Packet
in an intermediate structure and release the initial Packet when
there is 2 sides communication and treat the Packet that switched
the Flow to established.

Doing this limits the stress on the flow table that is only used when
we have a "complete" TCP session.

To avoid checking the Flow table to see if we have an established
Flow corresponding to the packet, we use a prefilter technique with
a hash table containing only the packet hash value. We don't need
to check for potential hash collision as the real lookup in the Flow
table will do its job later.
If the hash is not in the prefilter table then we have a new flow and
we can send the packet to the staging area.