applayer: incr flow counter if alproto for other dir was detected
Consider a case like the following.
-> Opposing direction data comes in first
-> At the beginning of Protocol Detection, the alproto for other side is
unknown
-> A successful call to TCPProtoDetectTriggerOpposingSide can however
lead to detection of alproto for the other side
-> Following this, a successful call is made to AppLayerParserParse

However, the flow counter is not incremented for this case.
This also does not map the number of flow events logged by Suricata for
such a case.

Increment the flow counter for a new flow for any applayer proto that is
accepted by that specific protocol's parser.

Related to Bug 7238