github-ci: test if bindgen bindings are up to date
This required us to remove the auto-generated by header as it can
change depending on what version is used.

I suppose different versions could generate slightly different output
that would cause this error out. In which case we may want to use a
custom script that can be a bit smarter output, or simply error out if
any of the headers used in the bindings is newer than the generated
_sys.rs file.