jsonbuilder: restrict visibility
JsonBuilder struct should not be made public because it is not
repr(C), and thus cannot be used as such by plugins,
because rust compiler does nots guarantee struct layout if we
do not use a fixed representation.

Also define a generic standard prototype for EveJsonSimpleTxLogFunc
and enfore it without cast for every function, each function
taking care of casting to its specific transaction type.