rust: fixes for breaking change on deranged crate
Deranged v0.4.1 (a dependency of the time crate) has implemented
PartialOrd for some integer types that conflict with the
implementation in the standard library creating an ambiguity as such
implementation are global. For more info see
https://github.com/jhpratt/deranged/issues/18.

To fix, use "as <type>" when we are comparing a one integer type to
another, to be fair the .into() here does seem odd.

And for our Direction type, use "from" instead of into.