util/lua: change arg name to reflect correctness
lua fns do not care about the argument count, they work with the index of
the item in the stack. Before library, there was just one item on the stack
so it worked out, however, with the library, the first item in the stack is
the libraray userdata, so, the fn fails because it expected string with
the existing hard coded setting of 1 for argc.