decode: expand to uint16_t; add documentation group
Expand to a uint16_t as there is a library use case to have some
user private flags, so reserve the last 4 flags for library use.

We were already at the max for a u8, so this opens up 4 more for our
use.

As #define's like this can't be logically grouped into an enum, try
Doxygen documentation groups, which create a group just of these flag
values and documents them together on a page.