modbus: trigger raw stream reassembly
Suricata's stream engine takes into account the chunk-size setting and
makes sure to trigger inspection on data when it reaches the chunk size.
However, this may be too late when it comes to inspection of smaller
protocol specific data which could result in incorrect data logged with
a transaction and logs misindicating the pkt that triggered an alert.

Fix this by making explicit call from all respective applayer parsers to
trigger raw stream reassembly to make data available for inspection in
the following call of data by the stream engine. This would ensure that
the inspection is not delayed and the correct data is logged per
transaction. This needs to happen per direction on the completion of an
entity like a request or a response. This only applies to applayer
protocols running over TCP.

Important notes:
1. The above mentioned behavior with and without this patch is
affected internally by the following conditions.
- inspection depth
- stream depth
In these special cases, the inspection window will be affected and
Suricata may not consider all the data that could be expected to be
inspected.
2. The inspection window is only considered up to the ACK'd data.

Modbus has a classic request response model, so, a call to trigger raw
stream reassembly is added on completion of each request and response.

Optimization 7026
Bug 7004