mdns: disable grouped logging by default
For mDNS, disable "grouped" logging by default. It adds a lot of data
with little value. The detailed format was kept as the default as it
applied to requests and responses, and no "grouped" formatting exists
for DNS requests.

Grouped can be enabled in the configuration file.

Also add flags to disable requests or responses.