(m)dns: logging configuration enhancements for mdns
Support more flexible DNS logging configuration to allow for mdns to
log in a less detailed format by default.

First we move part of the DNS logging configuration into Rust with
SCDnsLogConfig, this allows just a struct to be passed between C and
Rust instead of growing the function argument list.

Then we implement a new default logging configuration for mdns, which:

- Do not log additionals or authorities unless enabled

- Do not log answers in mdns requests by default, this is redundant
  information used for answer suppression.

Also expose in the mdns logging configuration the ability to
enable/disable request and response logging. By default, both are
logged.