dns: more log configuration flags
Make configuration flags for the following DNS fields:
- opcode
- flags
- id
- tx_id
- ttl
- rcode

These have less meaning in mDNS (and LLMNR) than in DNS, so it makes
sense to disable them in mdns logging.

If desired to have them as an option, rather than a user configurable
flag for each, maybe add a verbose flag to mdns logging.