ai: Agentic guidance file
Since VSCode Copilot offers a possibility to add an instruction
file to improve Agent results, this can be a good start informing
the AI agent about how development is done, what are the common commands,
it contains some suggested workflows.

This file can be helpful for any (agentic) AI, not just for VSCode.